+++ R/hpo_onsets_agg.R
@@ -7,6 +7,9 @@ hpo_onsets_agg <- function(hpo_onsets,
   if("hpo_id" %in% names(phenos)){
     hpo_onsets <- hpo_onsets[hpo_id %in% unique(phenos$hpo_id)]
   }
+  # f <- function(x, fun=mean){
+  #   if(all(is.na(x))) NA else fun(x,na.rm=TRUE)
+  # }
   annot_agg <- hpo_onsets[,.(onset=paste(unique(onset),collapse = ";"),
                         onset_names=paste(unique(onset_name),collapse = ";"),
                         onset_counts=paste(table(onset_name),collapse = ";"),
@@ -14,7 +17,7 @@ hpo_onsets_agg <- function(hpo_onsets,
                         onset_score_min=min(onset_score,na.rm=TRUE),
                         onset_score_max=max(onset_score,na.rm=TRUE)
   ),
-  by=agg_by]
+  by=agg_by] |> suppressWarnings()
   annot_agg$onset_top <- lapply(seq_len(nrow(annot_agg)),
                                 function(i){
                                   r <- annot_agg[i,]